IntegrateUpstreamModal: Factor out the modal
Factor out the modal for upstream integration into a dedicated component so that it can be shared easily
